requirements.txt
pdfplumber
pandas
openpyxl
customtkinter
Pillow
packaging